requirements.txt
beautifulsoup4==4.12.2
pycryptodome==3.15.0
PyMySQL==1.0.3
Requests==2.31.0
cryptography==41.0.4
paho-mqtt~=1.6.1